+++ packages/backend/src/socket/DataServer.test.ts
@@ -6,7 +6,7 @@ import { getPorts } from '../common/utils'
 test('start and stop data server', async () => {
   const ports = await getPorts()
   const dataServer = new DataServer(ports.dataServer)
-  expect(dataServer.io.engine.opts.cors).toBe(null) // No cors should be set by default
+  expect(dataServer.io.engine.opts.cors).toStrictEqual({}) // No cors should be set by default
   await dataServer.listen()
   await dataServer.close()
 })